refactor: restructure environment variable handling and remove deprecated files
