+++ src/lib/drizzle/database.ts
@@ -1,8 +1,7 @@
 import { drizzle } from 'drizzle-orm/node-postgres';
 import { Pool } from 'pg';
 
-import { env } from '@/env/server';
-
+import { env } from '../env';
 import * as schema from './schema';
 
 export const pool = new Pool({
